requirements.txt
boto3==1.35.69
Django==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-health-check==3.18.3
django-phonenumber-field==8.0.0
phonenumberslite==8.13.52
django-storages==1.14.4
django-ses==4.2.0
drf-yasg==1.21.7
psycopg2-binary==2.9.9
pillow==11.1.0
setuptools==74.1.2
requests==2.32.3
stripe==10.12.0
factory_boy==3.3.1
coverage==7.6.8
flake8==7.1.1
flake8-html==0.4.3
whitenoise==6.8.2
gunicorn==23.0.0